Disable source-highlight when CT_GDB_CROSS_STATIC_LIBSTDCXX is set
GDB does not allow using the source-highlight library, which is enabled
by default, when statically linking the libstdc++.

This commit updates the GDB build script to specify the
`--disable-source-highlight` option when CT_GDB_CROSS_STATIC_LIBSTDCXX
is enabled.

Note that this is a temporary workaround that is meant to address this
issue without adding direct support for the source-highlight library in
the crosstool-ng.

This commit should be reverted when source-highlight supported is added
to the crosstool-ng in the future and these conflicts can be resolved
in the Kconfig.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>